requirements.txt
asgi-redis==0.14.1
asgiref==3.2.10
autobahn==0.16.0
certifi==2020.6.20
cffi==1.14.3
channels==0.17.3
chardet==3.0.4
cryptography==3.4.7
daphne==0.15.0
defusedxml==0.6.0
dj-rest-auth==2.1.7
Django==3.1.2
django-allauth==0.43.0
django-cors-headers==3.5.0
django-rest-auth==0.9.5
djangorestframework==3.12.1
gunicorn==20.1.0
idna==2.10
msgpack-python==0.4.8
ndg-httpsclient==0.5.1
oauthlib==3.1.0
psycopg2-binary==2.8.6
pusher==3.0.0
pyasn1==0.4.8
pycparser==2.20
PyJWT==1.7.1
PyNaCl==1.4.0
pyOpenSSL==20.0.1
python3-openid==3.2.0
pytz==2020.1
redis==2.10.5
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
Twisted==16.4.1
txaio==2.5.1
urllib3==1.25.11
whitenoise==3.2.2
zope.interface==5.4.0
